requirements.txt
Flask==2.0.1
gluoncv==0.10.4.post1
ipywidgets==7.6.3
matplotlib==3.3.0
mxnet==1.7.0.post1
numpy
numpydoc
opencv-python==4.5.3.56
pandas==1.1
Pillow==8.3.1
pypaz==0.1.6
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.22.0
Werkzeug==2.0.1
zipp